changes on diff view & calculations /wip
