+++ src/frontend/package.json
@@ -14,6 +14,7 @@
     "axios": "^0.19.0",
     "balloon-css": "^1.0.4",
     "core-js": "^3.4.4",
+    "lodash.clonedeep": "^4.5.0",
     "moment": "^2.24.0",
     "tabulator-tables": "^4.5.3",
     "vue": "^2.6.10",
