+++ src/frontend/src/components/controllist/ControlList.scss
@@ -11,6 +11,12 @@
     margin-bottom:1px;
 }
 
+.control-CHANGED {
+    //background-color: red;
+    color: orange;
+}
+
+
 .control-REMOVED {
     //background-color: red;
     color: red;
