+++ src/frontend/src/services/MappingService.ts
@@ -1,196 +1,5 @@
 import { Check, CountersSummary, CheckSeverity, Collection } from '@/models'
-import { DiffCounters } from '@/models/ComponentDiff'
-
-export enum DiffOperation {
-    Added = 'ADDED',
-    Removed = 'REMOVED',
-    Changed = 'CHANGED',
-    Same = 'SAME'
-}
-
-export class CheckCollection {
-    score: number = 0
-    counters: CountersSummary = new CountersSummary()
-    objects: CheckObject[] = []
-    operation?: DiffOperation
-    empty: boolean = false;
-
-    constructor(public name: string, public type: string, public date: Date) {}
-
-    public static GetEmpty() : CheckCollection {
-        let empty = new CheckCollection("", "", new Date());
-        empty.empty = true;
-        return empty;
-    }
-    // compare objects
-    Compare(other: CheckCollection): [DiffOperation, DiffCounters] {
-
-        // parent Diff State
-        let operation = DiffOperation.Same;
-        let changes = new DiffCounters();
-
-        if(this.score !== other.score) {
-            operation = DiffOperation.Changed;
-        }
-
-        // compare this object array for removals
-        for(let i=0; i<this.objects.length;i++) {
-            let myObject = this.objects[i];
-            let otherIndex = other.objects.findIndex(x=>x.id === myObject.id);
-            if (otherIndex === -1) {
-                // this object is removed on other scan
-                operation = DiffOperation.Changed;                
-                myObject.operation = DiffOperation.Removed;
-                other.objects.push(CheckObject.GetEmpty(myObject.id))
-
-                changes.tick(myObject.operation)
-                continue;
-            }
-            // check object children
-            other.objects[otherIndex].Compare(myObject);           
-        }
-
-        // compare other object array for addition
-        for(let i=0; i<other.objects.length;i++) {
-            let otherObject = other.objects[i];
-            let myIndex = this.objects.findIndex(x=>x.id === otherObject.id);
-            if (myIndex === -1) {
-                // other obhect is removed from my scan
-                otherObject.operation = DiffOperation.Added
-                changes.tick(otherObject.operation)
-                operation = DiffOperation.Changed;
-                continue;
-            }
-            // check object children
-
-            this.objects[myIndex].Compare(otherObject);           
-            
-        }  
-        
-        return [operation, changes];
-    }
-}
-
-export class CheckObject {
-    id: string = ''
-    type: string = ''
-    name: string = ''
-    score: number = 0
-    counters: CountersSummary = new CountersSummary()
-    controls: CheckControl[] = []
-    controlGroups: CheckControlGroup[] = []
-    operation?: DiffOperation
-    checked: boolean = false;
-    empty: boolean = false;
-
-    public static GetEmpty(id:string) : CheckObject {
-        let empty = new CheckObject();
-        empty.id = id;
-        empty.empty = true;
-        return empty;
-    }
-
-    Compare(other: CheckObject): DiffCounters {
-        let changes = new DiffCounters();
-
-        // check controlGroups
-       
-        // check for removals
-        for(let j=0; j<this.controlGroups.length; j++) {
-            let myControlGroup = this.controlGroups[j];
-            let otherControlGroupIndex = other.controlGroups.findIndex(x=>x.name === myControlGroup.name);
-            if(otherControlGroupIndex === -1) {
-                this.operation = DiffOperation.Changed;
-                myControlGroup.operation = DiffOperation.Removed;
-                changes.tick(myControlGroup.operation);
-                continue;
-            }
-            let otherControlGroup = other.controlGroups[otherControlGroupIndex];
-            let [cgOperation, cgChanges] = otherControlGroup.Compare(myControlGroup);
-            myControlGroup.operation = cgOperation;
-            changes.merge(cgChanges);
-        }
-
-        // check for additions
-        for(let j=0; j<other.controlGroups.length; j++) {
-            let otherControlGroup = other.controlGroups[j];
-            let myControlGroupIndex = this.controlGroups.findIndex(x=>x.name === otherControlGroup.name);
-            if(myControlGroupIndex === -1) {
-                this.operation = DiffOperation.Changed;
-                otherControlGroup.operation = DiffOperation.Added;
-                changes.tick(otherControlGroup.operation);
-                continue;               
-            }
-            let myControlGroup = this.controlGroups[myControlGroupIndex];
-            let [cgOperation, cgChanges] = myControlGroup.Compare(otherControlGroup);
-            otherControlGroup.operation = cgOperation;
-            changes.merge(cgChanges);
-        }
-        
-        return changes;
-    }
-}
-
-export class CheckControl {
-    id: string = ''
-    text: string = ''
-    result: string = ''
-    icon: string = ''
-    order: number = 0
-    tags: string[] = []
-    operation?: DiffOperation
-}
-
-export class CheckControlGroup {
-    name: string = ''
-    items: CheckControl[] = []
-    operation?: DiffOperation
-
-    Compare(other: CheckControlGroup): [DiffOperation, DiffCounters] { 
-        let operation = DiffOperation.Same
-        let changes = new DiffCounters()
-
-        for(let i=0;i< this.items.length; i++) {
-            let myControl = this.items[i];
-            let otherControlIndex = other.items.findIndex(x=>x.id === myControl.id);
-            if(otherControlIndex === -1) {
-                myControl.operation = DiffOperation.Removed;
-                this.operation = DiffOperation.Changed;
-                changes.tick(this.operation);
-                continue;
-            }
-            let otherControl = other.items[otherControlIndex];            
-            if(myControl.result !== otherControl.result || myControl.text !== otherControl.text) {
-                myControl.operation = DiffOperation.Changed;
-                this.operation = DiffOperation.Changed;
-                changes.tick(this.operation);
-                continue;
-            }           
-            myControl.operation = DiffOperation.Same;
-        }
-
-        for(let i=0;i< other.items.length; i++) {
-            let otherControl = other.items[i];
-            let myControlIndex = this.items.findIndex(x=>x.id === otherControl.id);
-            if(myControlIndex === -1) {
-                otherControl.operation = DiffOperation.Added;
-                this.operation = DiffOperation.Changed;
-                changes.tick(this.operation);
-                continue;
-            }
-            let myControl = this.items[myControlIndex];            
-            if(myControl.result !== otherControl.result || myControl.text !== otherControl.text) {
-                otherControl.operation = DiffOperation.Changed;
-                this.operation = DiffOperation.Changed;
-                changes.tick(this.operation);
-                continue;
-            }
-            myControl.operation = DiffOperation.Same;
-        }
-
-        return [operation, changes];
-    }
-}
+import { CheckCollection, CheckControl, CheckControlGroup, CheckObject } from './DiffService'
 
 export class MappingService {
     public static getResultsByCategory(checks: Check[]): any[] {
